+++ shared/azure-cosmosdb-data-to-mongodb.ts
@@ -1,4 +1,4 @@
-import { SASProtocol } from "@azure/storage-blob";
+import { SASProtocol } from '@azure/storage-blob';
 import {
   MongoClient,
   Filter,
@@ -7,8 +7,8 @@ import {
   SortDirection,
   FindOptions,
   Sort,
-  IndexSpecification,
-} from "mongodb";
+  IndexSpecification
+} from 'mongodb';
 
 export async function uploadArrToMongoDb(
   connectionString: string,
@@ -18,13 +18,13 @@ export async function uploadArrToMongoDb(
 ): Promise<InsertManyResult<Document>> {
   // Data exists and looks like its in the correct shape
 
-  if (!connectionString) throw new Error("connection string if missing");
-  if (!databaseName) throw new Error("databaseName is missing");
-  if (!collectionName) throw new Error("collectionName is missing");
+  if (!connectionString) throw new Error('connection string if missing');
+  if (!databaseName) throw new Error('databaseName is missing');
+  if (!collectionName) throw new Error('collectionName is missing');
 
-  if (!arrData) throw new Error("data is missing");
+  if (!arrData) throw new Error('data is missing');
   if (!arrData?.length || arrData.length === 0)
-    throw new Error("data found to be empty");
+    throw new Error('data found to be empty');
 
   console.log(connectionString);
   console.log(databaseName);
@@ -34,13 +34,13 @@ export async function uploadArrToMongoDb(
 
   // Use connect method to connect to the server
   await client.connect();
-  console.log("Connected successfully to MongoDB");
+  console.log('Connected successfully to MongoDB');
 
   const insertResult = await client
     .db(databaseName)
     .collection(collectionName)
     .insertMany(arrData);
-  
+
   client.close();
   return insertResult;
 }
@@ -52,23 +52,23 @@ export async function findInMongoDb(
   query: Filter<Document>,
   options: any
 ): Promise<Array<unknown>> {
-  if (!connectionString) throw new Error("connection string if missing");
-  if (!databaseName) throw new Error("databaseName is missing");
-  if (!collectionName) throw new Error("collectionName is missing");
+  if (!connectionString) throw new Error('connection string if missing');
+  if (!databaseName) throw new Error('databaseName is missing');
+  if (!collectionName) throw new Error('collectionName is missing');
 
-  if (!query) throw new Error("query is missing");
+  if (!query) throw new Error('query is missing');
 
   const client = new MongoClient(connectionString);
 
   // Use connect method to connect to the server
   await client.connect();
-  console.log("Connected successfully to MongoDB");
+  console.log('Connected successfully to MongoDB');
 
   // const sort = {};
   // sort["updatedAt"] = sortDirection;
 
   const options2 = {
-    sort: { updatedAt: -1 },
+    sort: { updatedAt: -1 }
   };
 
   const findResult = await client
@@ -85,35 +85,35 @@ export async function findRepoInMongoDbWithProjection(
   collectionName: string,
   repoName: string
 ): Promise<Array<unknown>> {
-  if (!connectionString) throw new Error("connection string if missing");
-  if (!databaseName) throw new Error("databaseName is missing");
-  if (!collectionName) throw new Error("collectionName is missing");
+  if (!connectionString) throw new Error('connection string if missing');
+  if (!databaseName) throw new Error('databaseName is missing');
+  if (!collectionName) throw new Error('collectionName is missing');
 
   const client = new MongoClient(connectionString);
   await client.connect();
   const query = {
     repositoryName: {
-      $eq: repoName,
-    },
+      $eq: repoName
+    }
   };
 
   var projection = {
     customDateUploaded: 1.0,
-    "stargazers.totalCount": 1.0,
-    "forks.totalCount": 1.0,
-    "issues.totalCount": 1.0,
-    "pullRequests.totalCount": 1.0,
-    _id: 0.0,
+    'stargazers.totalCount': 1.0,
+    'forks.totalCount': 1.0,
+    'issues.totalCount': 1.0,
+    'pullRequests.totalCount': 1.0,
+    _id: 0.0
   };
 
   const sort: Sort = {
-    customDateUploaded: -1,
+    customDateUploaded: -1
   };
 
   var options: FindOptions = {
     sort,
     projection,
-    limit: 100,
+    limit: 100
   };
 
   // @ts-ignore
@@ -202,19 +202,23 @@ export async function findOrgReposInMongoDbWithProjection(
   databaseName: string,
   collectionName: string
 ): Promise<Array<unknown>> {
-  if (!connectionString) throw new Error("connection string if missing");
-  if (!databaseName) throw new Error("databaseName is missing");
-  if (!collectionName) throw new Error("collectionName is missing");
+  if (!connectionString) throw new Error('connection string if missing');
+  if (!databaseName) throw new Error('databaseName is missing');
+  if (!collectionName) throw new Error('collectionName is missing');
 
   const client = new MongoClient(connectionString);
   await client.connect();
 
-  const dateAfter = await findDateGreaterThan(connectionString, databaseName, collectionName+"-count");
+  const dateAfter = await findDateGreaterThan(
+    connectionString,
+    databaseName,
+    collectionName + '-count'
+  );
 
-    const query = {
-      customDateUploaded: {$gt: dateAfter}
-    };
-  
+  const query = {
+    customDateUploaded: { $gt: dateAfter }
+  };
+  /*
     var projection = {
       customDateUploaded: 1.0,
       repositoryName: 1.0,
@@ -224,53 +228,70 @@ export async function findOrgReposInMongoDbWithProjection(
       "issues.totalCount": 1.0,
       "pullRequests.totalCount": 1.0,
       _id: 0.0,
-    };
-  
-    const sort: Sort = {
-       customDateUploaded: -1
-     };
-  
-    var options: FindOptions = {
-      sort,
-      projection
-    };
-  
-    // // @ts-ignore
-    var docs = await client
-      .db(databaseName)
-      .collection(collectionName)
-      .find(query, options)
-      .toArray();
-  
-    // transform
-    docs.map((doc) => {
-      doc.stars = doc.stargazers.totalCount;
-      doc.forks = doc.forks.totalCount;
-      doc.issues = doc.issues.totalCount;
-      doc.pr = doc.pullRequests.totalCount;
-  
-      delete doc.stargazers;
-      delete doc.pullRequests;
-      delete doc.forks.totalCount;
-      delete doc.issues.totalCount;
-    });
-  
-    docs.sort()
+    };*/
 
-    client.close();
+  var projection = {
+    lastPushToDefaultBranch: 1.0,
+    is: 1.0,
+    has: 1.0,
+    legal: 1.0,
+    languages: 1.0,
+    diskUsage: 1.0,
+    date: 1.0,
+    customDateUploaded: 1,
+    repositoryName: 1.0,
+    'watchers.totalCount': 1,
+    'stargazers.totalCount': 1.0,
+    'forks.totalCount': 1.0,
+    'issues.totalCount': 1.0,
+    'pullRequests.totalCount': 1.0
+  };
 
-    docs.sort(function (a, b) {
+  const sort: Sort = {
+    customDateUploaded: -1
+  };
 
-      var nameA = a.repositoryName.toLowerCase(), nameB = b.repositoryName.toLowerCase()
-  
-      if (nameA < nameB)
-        return -1;
-      if (nameA > nameB)
-        return 1;
-      return 0;  // No sorting
-  })
+  var options: FindOptions = {
+    sort,
+    projection
+  };
 
-    return docs;
+  // // @ts-ignore
+  var docs = await client
+    .db(databaseName)
+    .collection(collectionName)
+    .find(query, options)
+    .toArray();
+
+  // transform
+  docs.map((doc) => {
+    doc.watchers = doc.watchers.totalCount;
+    doc.stars = doc.stargazers.totalCount;
+    doc.forks = doc.forks.totalCount;
+    doc.issues = doc.issues.totalCount;
+    doc.pr = doc.pullRequests.totalCount;
+
+    delete doc.watchers.totalCount;
+    delete doc.stargazers;
+    delete doc.pullRequests;
+    delete doc.forks.totalCount;
+    delete doc.issues.totalCount;
+  });
+
+  docs.sort();
+
+  client.close();
+
+  docs.sort(function (a, b) {
+    var nameA = a.repositoryName.toLowerCase(),
+      nameB = b.repositoryName.toLowerCase();
+
+    if (nameA < nameB) return -1;
+    if (nameA > nameB) return 1;
+    return 0; // No sorting
+  });
+
+  return docs;
 }
 
 export async function findDateGreaterThan(
@@ -278,9 +299,9 @@ export async function findDateGreaterThan(
   databaseName: string,
   collectionName: string
 ): Promise<unknown> {
-  if (!connectionString) throw new Error("connection string if missing");
-  if (!databaseName) throw new Error("databaseName is missing");
-  if (!collectionName) throw new Error("collectionName is missing");
+  if (!connectionString) throw new Error('connection string if missing');
+  if (!databaseName) throw new Error('databaseName is missing');
+  if (!collectionName) throw new Error('collectionName is missing');
 
   console.log(`${databaseName}/${collectionName}`);
 
@@ -288,7 +309,7 @@ export async function findDateGreaterThan(
   await client.connect();
   const query = {};
   const sort: Sort = {
-    date: -1,
+    date: -1
   };
   var projection = {
     date: 1.0,
@@ -300,8 +321,6 @@ export async function findDateGreaterThan(
     limit: 1,
     projection
   };
-
-
 
   // @ts-ignore
   var docs = await client
@@ -316,25 +335,25 @@ export async function findDateGreaterThan(
 export async function findSummaryInMongoDb(
   connectionString: string,
   databaseName: string,
-  collectionName: string,
+  collectionName: string
 ): Promise<Array<unknown>> {
-  if (!connectionString) throw new Error("connection string if missing");
-  if (!databaseName) throw new Error("databaseName is missing");
-  if (!collectionName) throw new Error("collectionName is missing");
+  if (!connectionString) throw new Error('connection string if missing');
+  if (!databaseName) throw new Error('databaseName is missing');
+  if (!collectionName) throw new Error('collectionName is missing');
 
   const client = new MongoClient(connectionString);
 
   // Use connect method to connect to the server
   await client.connect();
-  console.log("Connected successfully to MongoDB");
+  console.log('Connected successfully to MongoDB');
 
   const query = {};
   const sort: Sort = {
-    date: -1,
+    date: -1
   };
   var projection = {
     date: 1.0,
-    count:1.0,
+    count: 1.0,
     _id: 0.0
   };
   var options: FindOptions = {
@@ -346,7 +365,8 @@ export async function findSummaryInMongoDb(
   const findResult = await client
     .db(databaseName)
     .collection(collectionName)
-    .find(query, options).toArray();
+    .find(query, options)
+    .toArray();
 
   return findResult;
 }